requirements.txt
python-binance==1.0.19
pandas==2.1.4
pandas-ta==0.3.14b0
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.7.0
schedule==1.2.1
pytest==7.4.4
pytest-mock==3.12.0
